// web/src/app/models/pipeline.ts
/**
 * High Five
 * The OpenAPI specification for High Five's controllers
 *
 * OpenAPI spec version: 0.0.1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Pipeline {
  name?: string;
  /**
   * Unique id associated with the pipeline
   */
  id?: string;
  metadataType?: string;
  tools?: Array<string>;
}